move worker parse_stdout_* fns to serialize_util
