+++ depobs/website/config.py
@@ -24,10 +24,11 @@ LOGGING = {
         "request.summary": {"handlers": ["console"], "level": "INFO"},
         "depobs.clients.cratesio": {"handlers": ["console"], "level": "INFO"},
         "depobs.clients.github": {"handlers": ["console"], "level": "INFO"},
-        "depobs.clients.npmsio": {"handlers": ["console"], "level": "INFO"},
         "depobs.clients.npm_registry": {"handlers": ["console"], "level": "INFO",},
+        "depobs.clients.npmsio": {"handlers": ["console"], "level": "INFO"},
         "depobs.database.models": {"handlers": ["console"], "level": "INFO"},
         "depobs.database.serializers": {"handlers": ["console"], "level": "INFO"},
+        "depobs.util.serialize_util": {"handlers": ["console"], "level": "INFO"},
         "depobs.website.views": {"handlers": ["console"], "level": "INFO"},
         "depobs.worker.k8s": {"handlers": ["console"], "level": "INFO",},
         "depobs.worker.scoring": {"handlers": ["console"], "level": "INFO",},
